feedmaker/settings/local.py
import os
from dotenv import load_dotenv, find_dotenv
from feedmaker.settings.base import *

load_dotenv(find_dotenv())
SECRET_KEY = os.environ["SECRET_KEY"]

DEBUG = True

MIDDLEWARE += [
    "django.middleware.cache.UpdateCacheMiddleware",
    "django.middleware.common.CommonMiddleware",
    "django.middleware.cache.FetchFromCacheMiddleware",
]

CACHES = {
    "default": {
        #'BACKEND': 'django.core.cache.backends.filebased.FileBasedCache',
        #'LOCATION': '/tmp/feedmaker/',
        "BACKEND": "django.core.cache.backends.dummy.DummyCache",
    }
}

CACHE_MIDDLEWARE_ALIAS = "default"
CACHE_MIDDLEWARE_SECONDS = 60 * 5
CACHE_MIDDLEWARE_KEY_PREFIX = ""

LOGGING = {
    "version": 1,
    "disable_existing_loggers": False,
    "handlers": {
        "console": {
            "class": "logging.StreamHandler",
        },
    },
    "root": {
        "handlers": ["console"],
        "level": "INFO",
    },
}
