requirements.txt
appdirs==1.4.4
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.11.2
black==23.1.0
bs4==0.0.1
certifi==2022.12.7
cfgv==3.3.1
charset-normalizer==3.1.0
click==8.1.3
cssselect==1.2.0
distlib==0.3.6
Django==4.1.7
extruct==0.14.0
fake-useragent==1.1.1
filelock==3.9.0
gunicorn==20.1.0
hiredis==2.2.2
html-text==0.5.2
html5lib==1.1
identify==2.5.20
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
isodate==0.6.1
jstyleson==0.0.2
lxml==4.9.2
mf2py==1.1.2
mypy-extensions==1.0.0
nodeenv==1.7.0
packaging==23.0
parse==1.19.0
pathspec==0.11.0
platformdirs==3.1.1
pre-commit==3.1.1
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
pyRdfa3==3.5.3
python-dotenv==1.0.0
PyYAML==6.0
rdflib==6.2.0
redis==4.5.1
requests==2.28.2
requests-file==1.5.1
requests-html==0.10.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
virtualenv==20.21.0
w3lib==2.1.1
webencodings==0.5.1
websockets==10.4
zipp==3.15.0
